fix warning separate ans

+++ templates/main.cpp
@@ -49,7 +49,7 @@ ll powMod(ll a, ll b){if(!b)return 1;ll r=powMod(a,b/2);r=mulMod(r,r);return(b%2
 
 template <typename T>
 ostream& operator<<(ostream& os, const vector<T>& v) {
-	for (ll i=0;i<v.size();i++) os<<v[i]<<" \n"[i+1==v.size()];
+	for (size_t i=0;i<v.size();i++) os<<v[i]<<" \n"[i+1==v.size()];
 	return os;
 }
 template <typename T>
